requirements.txt
transformers
torch
requests
beautifulsoup4
newspaper3k
rapidfuzz
